Move Contract into its own file
